Complete token authentication update

+++ backend/routes/posts.routes.js
@@ -15,7 +15,7 @@ const MIME_TYPE_MAP = {
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {
         const isValid = MIME_TYPE_MAP[file.mimetype];
-        let error = new Error('Invalid mime type');
+        let error = new Error(`Invalid mime type`);
         if(isValid) {
             error = null;
         }
@@ -46,7 +46,7 @@ router.get('/', (req, res, next) => {
     })
     .then((count) => {
         res.status(200).json({
-            message: "Posts fetched succesfully!",
+            message: `Posts fetched succesfully!`,
             posts: fetchedPosts,
             postCounts: count
         });
@@ -62,7 +62,7 @@ router.get('/:id', (req, res, next) => {
                 post: post
             })
         } else {
-            res.status(404).json({message: 'Post not found'});
+            res.status(404).json({message: `Post not found`});
         }
     })
 });
@@ -74,11 +74,12 @@ router.post('/', checkAuth, upload.single('image'), (req, res, next) => {
         email: req.body.email,
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/uploads/' + req.file.filename
+        imagePath: url + '/uploads/' + req.file.filename,
+        creatorID: req.userData.userID
     });
     post.save().then((result) => {
         res.status(201).json({
-            message: 'Post added successfully!',
+            message: `Post added successfully!`,
             post: {
                 ...result,
                 id: result._id
@@ -104,17 +105,24 @@ router.put('/:id', checkAuth, upload.single('image'), (req, res, next) => {
         const url = req.protocol + '://' + req.get('host');
         imagePathnew = url + '/uploads/' + req.file.filename;
     };
-    const post = {
+    const post = new Post({
         name: req.body.name,
         email: req.body.email,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePathnew
-    };
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        res.status(200).json({
-            message: "Update successfull"
-        });
+        imagePath: imagePathnew,
+        creatorID: req.userData.userID
+    });
+    Post.updateOne({_id: req.params.id, creatorID: req.userData.userID}, post).then(result => {
+        if(result.nModified > 0) {
+            res.status(200).json({
+                message: `Post ${post.name} updated successfully.`
+            });
+        } else {
+            res.status(401).json({
+                message: `Not authorized to update.`
+            });
+        }   
     });
 });
 
@@ -126,16 +134,22 @@ router.delete('/:id', checkAuth, (req, res, next) => {
             return s.substr(prefix.length);
         }
         imagePath = removePrefix('http:/localhost:3030/uploads//',imagePath);
-        var query = Post.deleteOne({_id: req.params.id});
+        var query = Post.deleteOne({_id: req.params.id, creatorID: req.userData.userID});
         var promise = query.exec();
-        promise.then(() => {
-            fs.unlink('F:\\MEAN\\Project with Angular 2+\\Project1\\backend\\uploads\\' + imagePath, (err) => {
-                if (err) throw err;
-                console.log(imagePath + ' was deleted');
-            });
-            res.status(201).json({
-            message: `Post deleted successfully!`
-            });
+        promise.then((result) => {
+            if(result.n > 0) {
+                fs.unlink('F:\\MEAN\\Project with Angular 2+\\Project1\\backend\\uploads\\' + imagePath, (err) => {
+                    if (err) throw err;
+                    console.log(`${imagePath} deleted.`);
+                });
+                res.status(201).json({
+                    message: `Post deleted successfully!`
+                });
+            } else {
+                res.status(401).json({
+                    message: `Not authorized to update.`
+                })
+            }
         });
     });
 });
